requirements.txt
aiohttp==3.8.4
python-telegram-bot==20.6
python-dotenv==1.0.0
